+++ backend/internal/handlers/github/routes.go
@@ -9,16 +9,17 @@ import (
 func Routes(app *fiber.App, params types.Params) {
 	service := newGitHubService(params.Store, params.GitHubApp, &params.UserCfg)
 
+	app.Get("/callback", service.GetCallbackURL())
 	app.Post("/login", service.Login())
 	app.Post("/logout", service.Logout())
 
 	protected := app.Group("/github")
 	protected.Use(middleware.Protected(params.UserCfg.JWTSecret))
 	protected.Get("/user", service.GetCurrentUser())
-  
-  protected.Post("/sync", service.SyncAssignments)
 
-  protected.Get("/classrooms", service.ListClassrooms())
+	protected.Post("/sync", service.SyncAssignments)
+
+	protected.Get("/classrooms", service.ListClassrooms())
 	protected.Post("/startup", service.AppInitialization())
 
 	protected.Get("/orgs/:org", service.GetOrg())
