+++ backend/internal/models/organization.go
@@ -2,6 +2,11 @@ package models
 
 import "time"
 
+type RolesResponse struct {
+	TotalCount int                `json:"total_count"`
+	Roles      []OrganizationRole `json:"roles"`
+}
+
 type OrganizationTemplateRole struct {
 	Name        string
 	Description string
@@ -21,15 +26,9 @@ type OrganizationRole struct {
 
 type Organization struct {
 	SimpleOrganization
-	// Name              *string `json:"name,omitempty"`
-	Email *string `json:"email,omitempty"`
-	// Login             string  `json:"login"`
-	// ID                int64   `json:"id"`
-	// NodeID            string  `json:"node_id"`
-	// AvatarURL         string  `json:"avatar_url"`
-	GravatarID *string `json:"gravatar_id,omitempty"`
-	URL        string  `json:"url"`
-	// HTMLURL           string  `json:"html_url"`
+	Email             *string `json:"email,omitempty"`
+	GravatarID        *string `json:"gravatar_id,omitempty"`
+	URL               string  `json:"url"`
 	FollowersURL      string  `json:"followers_url"`
 	FollowingURL      string  `json:"following_url"`
 	GistsURL          string  `json:"gists_url"`
@@ -52,40 +51,3 @@ type SimpleOrganization struct {
 	Name      string `json:"name"`
 	AvatarURL string `json:"avatar_url"`
 }
-
-var Prof_Role = OrganizationTemplateRole{
-	Name:        "Professor",
-	Description: "Professor",
-	Permissions: []string{},
-	BaseRole:    "admin",
-}
-
-var TA_Role = OrganizationTemplateRole{
-	Name:        "TA",
-	Description: "Teaching Assistant",
-	Permissions: []string{},
-	BaseRole:    "maintain",
-}
-
-var Inactive_TA_Role = OrganizationTemplateRole{
-	Name:        "Inactive TA",
-	Description: "Inactive Teaching Assistant",
-	Permissions: []string{},
-	BaseRole:    "read",
-}
-
-var Student_Role = OrganizationTemplateRole{
-	Name:        "Student",
-	Description: "Student",
-	Permissions: []string{},
-	BaseRole:    "read",
-}
-
-var Inactive_Student_Role = OrganizationTemplateRole{
-	Name:        "Inactive Student",
-	Description: "Inactive Student",
-	Permissions: []string{},
-	BaseRole:    "read",
-}
-
-var AllRoles = []OrganizationTemplateRole{Prof_Role, TA_Role, Inactive_TA_Role, Student_Role, Inactive_Student_Role}
